# Improve performance of uploads to GCP bucket